complete

+++ microservices/booking-service/src/main/java/com/zosh/modal/SaloonReport.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class SaloonReport {
    private Long saloonID;
    private String saloonName;
-   private Double totalEarning;
+   private int totalEarning;
    private Integer totalBooking;
    private Integer cancelledBooking;
    private Double totalRefund;
